feat : 민감정보 필터링 메소드 util로 분리
